Saving a ChannelHandler
